Add clean directory call randomly
