+++ services/search/src/search/config.py
@@ -18,6 +18,10 @@ from libcommon.config import (
 )
 
 DUCKDB_INDEX_CACHE_DIRECTORY = None
+DUCKDB_INDEX_CACHE_CLEAN_CACHE_PROBA = 0.05
+DUCKDB_INDEX_CACHE_MAX_NUMBER_OF_FOLDERS_TO_CLEAN = 50
+DUCKDB_INDEX_CACHE_EXPIRED_TIME_INTERVAL_SECONDS = 3_600  # 1 hour
+
 DUCKDB_INDEX_TARGET_REVISION = "refs/convert/duckdb"
 DUCKDB_INDEX_EXTENSIONS_DIRECTORY: Optional[str] = None
 
@@ -25,6 +29,8 @@ DUCKDB_INDEX_EXTENSIONS_DIRECTORY: Optional[str] = None
 @dataclass(frozen=True)
 class DuckDbIndexConfig:
     cache_directory: Optional[str] = DUCKDB_INDEX_CACHE_DIRECTORY
+    clean_cache_proba: float = DUCKDB_INDEX_CACHE_CLEAN_CACHE_PROBA
+    expired_time_interval_seconds: int = DUCKDB_INDEX_CACHE_EXPIRED_TIME_INTERVAL_SECONDS
     target_revision: str = DUCKDB_INDEX_TARGET_REVISION
     extensions_directory: Optional[str] = DUCKDB_INDEX_EXTENSIONS_DIRECTORY
 
@@ -34,6 +40,10 @@ class DuckDbIndexConfig:
         with env.prefixed("DUCKDB_INDEX_"):
             return cls(
                 cache_directory=env.str(name="CACHE_DIRECTORY", default=DUCKDB_INDEX_CACHE_DIRECTORY),
+                clean_cache_proba=env.float(name="CLEAN_CACHE_PROBA", default=DUCKDB_INDEX_CACHE_CLEAN_CACHE_PROBA),
+                expired_time_interval_seconds=env.int(
+                    name="EXPIRED_TIME_INTERVAL_SECONDS", default=DUCKDB_INDEX_CACHE_EXPIRED_TIME_INTERVAL_SECONDS
+                ),
                 target_revision=env.str(name="TARGET_REVISION", default=DUCKDB_INDEX_TARGET_REVISION),
                 extensions_directory=env.str(name="EXTENSIONS_DIRECTORY", default=DUCKDB_INDEX_EXTENSIONS_DIRECTORY),
             )
